refactor(electron): inline bruno-pool and bruno-storage into electron services
Nothing outside electron consumed either package, so they live inside
bruno-electron now as siblings of mount.

- services/pool/index.js: Pool, getPool, destroyPool, JobType (was pool.ts)
- services/pool/worker.js: workerpool.worker(...) registration (was worker.ts)
- services/pool/jobs/parse-file.js: the parse-file job (was scripts/parse-file.ts)
  Drops the unreachable type === 'dotenv' branch.
- services/storage/index.js: lazy node:sqlite wrapper with migrations,
  prepared-statement cache, transaction() — unchanged in behavior.
- services/mount/manager.js and services/mount/gitlite.js require the new
  local services instead of @usebruno/pool / @usebruno/storage.
- packages/bruno-electron/package.json drops the two @usebruno deps and
  adds workerpool@10.0.2 as a direct dep.
- Root package.json, scripts/setup.js, CI workflows, .gitignore drop the
  workspace and build entries for the removed packages.
