+++ .github/actions/common/setup-node-deps/action.yml
@@ -35,5 +35,3 @@ runs:
         npm run build:bruno-requests
         npm run build:schema-types
         npm run build:bruno-filestore
-        npm run build:bruno-pool
-        npm run build:bruno-storage
