+++ .github/workflows/flaky-test-detector.yml
@@ -56,8 +56,6 @@ jobs:
           npm run build:bruno-requests
           npm run build:schema-types
           npm run build:bruno-filestore
-          npm run build:bruno-pool
-          npm run build:bruno-storage
 
       - name: Run Playwright tests
         run: xvfb-run npm run test:e2e
