+++ .gitignore
@@ -72,5 +72,3 @@ packages/bruno-filestore/dist
 packages/bruno-requests/dist
 packages/bruno-schema-types/dist
 packages/bruno-converters/dist
-packages/bruno-pool/dist
-packages/bruno-storage/dist
