+++ package.json
@@ -16,9 +16,7 @@
     "packages/bruno-toml",
     "packages/bruno-graphql-docs",
     "packages/bruno-requests",
-    "packages/bruno-filestore",
-    "packages/bruno-pool",
-    "packages/bruno-storage"
+    "packages/bruno-filestore"
   ],
   "homepage": "https://usebruno.com",
   "devDependencies": {
@@ -68,8 +66,6 @@
     "build:bruno-common": "npm run build --workspace=packages/bruno-common",
     "build:bruno-requests": "npm run build --workspace=packages/bruno-requests",
     "build:bruno-filestore": "npm run build --workspace=packages/bruno-filestore",
-    "build:bruno-pool": "npm run build --workspace=packages/bruno-pool",
-    "build:bruno-storage": "npm run build --workspace=packages/bruno-storage",
     "build:bruno-converters": "npm run build --workspace=packages/bruno-converters",
     "build:bruno-query": "npm run build --workspace=packages/bruno-query",
     "build:graphql-docs": "npm run build --workspace=packages/bruno-graphql-docs",
