+++ packages/bruno-electron/package.json
@@ -40,12 +40,10 @@
     "@usebruno/converters": "^0.1.0",
     "@usebruno/filestore": "^0.1.0",
     "@usebruno/js": "0.12.0",
-    "@usebruno/pool": "0.1.0",
     "@usebruno/lang": "0.12.0",
     "@usebruno/node-machine-id": "^2.0.0",
     "@usebruno/requests": "^0.1.0",
     "@usebruno/schema": "0.7.0",
-    "@usebruno/storage": "0.1.0",
     "about-window": "^1.15.2",
     "adm-zip": "^0.5.16",
     "ai": "6.0.39",
@@ -81,6 +79,7 @@
     "socks-proxy-agent": "^8.0.2",
     "tough-cookie": "^6.0.0",
     "uuid": "^10.0.0",
+    "workerpool": "10.0.2",
     "yup": "^0.32.11",
     "zod": "^4.1.8"
   },
