+++ packages/bruno-electron/src/services/pool/worker.js
@@ -1,0 +1,7 @@
+const workerpool = require('workerpool');
+const parseFile = require('./jobs/parse-file');
+const { JobType } = require('./index');
+
+workerpool.worker({
+  [JobType.ParseFile]: parseFile
+});
