--- packages/bruno-pool/src/worker.ts
@@ -1,7 +1,0 @@
-import * as workerpool from 'workerpool';
-import parseFile from './scripts/parse-file';
-import { JobType } from './pool';
-
-workerpool.worker({
-  [JobType.ParseFile]: parseFile
-});
