--- packages/bruno-storage/src/index.ts
@@ -1,2 +1,0 @@
-export { Database } from './database';
-export type { DatabaseOptions, Migration, PathLike } from './database';
