+++ scripts/setup.js
@@ -98,8 +98,6 @@ async function setup() {
     execCommand('npm run build:bruno-requests', 'Building bruno-requests');
     execCommand('npm run build:schema-types', 'Building schema-types');
     execCommand('npm run build:bruno-filestore', 'Building bruno-filestore');
-    execCommand('npm run build:bruno-pool', 'Building bruno-pool');
-    execCommand('npm run build:bruno-storage', 'Building bruno-storage');
 
     // Bundle JS sandbox libraries
     execCommand('npm run sandbox:bundle-libraries --workspace=packages/bruno-js', 'Bundling JS sandbox libraries');
